requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv>=1.0.0
pandas==2.1.3
python-dateutil==2.8.2
openpyxl==3.1.2
python-calamine>=0.1
requests==2.31.0
pyodbc>=4.0.0
msal